feat: Add expression evaluation back to REPL (#18)
This is chapter 8 challenge 1. This is a case where using a class for
the parser would be helpful, since it would be easier to make both the
`parse` and `expression` methods public.

I'm following the Node readline tiny CLI example:
https://nodejs.org/api/readline.html#example-tiny-cli

It's nice to have up arrow in the CLI, but there's definitely a flicker
that I don't love.