feat: Implement Chapter 10 (Functions) (#21)
Implementation of [Chapter 10] of _Crafting Interpreters_.

I wondered back in #10 whether I'd be shooting myself in the foot by
supporting `,` as a number separator (`1,234`). This is where it
happened. The comma becomes quite ambiguous in the arguments list of a
function call. For exmaple:

```
f(1,234,567);
```

Is that a call to `f` with a single argument, `1,234,567` or a call to
`f` with three arguments, `1`, `234` and `567`? The grammar is
ambiguous.

My solution is to make spaces significant while scanning:

```
f(1,234,567);  // one argument
f(1, 234, 567);  // three arguments
```

We'll see how this holds up!

[Chapter 10]: https://craftinginterpreters.com/functions.html