feat: Implement Chapter 11 (Resolving and Binding) (#22)
Implementation of _Crafting Interpreters_ [Chapter 11: Resolving and
Binding](https://craftinginterpreters.com/resolving-and-binding.html).

This introduces a variable resolution pass with the goal of making sure
that this code:

```
var a = "global";
{
  fun showA() {
    print a;
  }

  showA();
  var a = "block";
  showA();
}
```

prints global / global instead of global / block.

This is a fairly straightforward translation of the Java code in the
book. I wound up going with the closure approach rather than the class
approach to avoid having to write type annotations on all the visitor
methods.

Storing the number of hops you need to go up the environment chain seems
like an incredibly brittle approach, but that's how the book does it. My
inclination would have been to create a new scope with every `var`
statement.

Other notable things:

- I wrote a baseline updater script. This runs in Node and is much
faster than the shell script.
- The updater and end-to-end test both use `fs.readdir` with
`{recursive:true}`, which requires Node.js 20+.