chore: Refactor to make Interpreter not be a visitor (#35)
After [puzzling
over](https://stackoverflow.com/questions/77876338/could-typescript-model-this-function-if-it-had-dependent-types)
the visitor pattern in TypeScript, @mulias pointed out that I didn't
really need it; pattern matching (via a switch statement) is the more
idiomatic way to do this with a tagged union.

This is that refactor. There's a lot fewer imports, at least. No visitor
but we still get exhaustiveness checking. For the `evaluate` method it's
via the implicit return. For the `execute` method on expressions, we
have to bake it in more explicitly.